\name{InitErgmTerms}
\alias{InitErgmTerms}
\alias{ergm.userterms}
\alias{InitErgmTerm.mindegree}
\alias{InitErgmTerm.nmedges}
\alias{InitErgmTerm.nmtriangle}
\alias{InitErgmTerm.nmkstar}
\alias{InitErgmTerm.blkedges}
\alias{InitErgmTerm.blktriangle}
\alias{InitErgmTerm.blkkstar}
\alias{InitErgmTerm.blktriangle2}
\alias{InitErgmTerm.blkkstar2}
\alias{InitErgmTerm.blktriangle3}
\alias{InitErgmTerm.blkkstar3}
\alias{InitErgmTerm.degseq}
\alias{InitErgmTerm.blkdegseq}
\docType{package}
\title{
User-defined terms used in Exponential Family Random Graph Models
}
\description{
These functions are required for the interface between ergm package and ergm.userterms package.These functions are written for the purpose of initialize the terms in particular models.See \link{blkergm-terms} for more information.
}
\references{
Hunter DR, Goodreau SM, Handcock MS (2011)
{\pkg{ergm.userterms}: {A} Template Package for 
Extending \pkg{statnet}}.
}
\seealso{statnet, network, ergm, ergm-terms}
\keyword{ package }
\keyword{ models }
