\name{zachary}
\docType{data}
\title{Zachary Karate Club social network data}
\alias{zachary}
\alias{zachary.club}
\description{
\link{zachary.club} is Social network of friendships between 34 members of a karate club at a US university in the 1970. It is an undirected network without attributes of vertices.
}

\usage{data(zachary)}
\format{
\code{zachary} is a \code{\link[network]{network}} object with 34 vertices 
and 78 undirected edges.  To obtain additional 
summary information about it,
type \code{summary(zachary)}.
}

\seealso{\code{\link[network]{network}}, \code{\link[network]{plot.network}}, 
\code{\link{ergm}}, \code{\link{zachary.club}}}


\references{
W. W. Zachary. \emph{An information flow model for conflict and fission in small groups}, \emph{Journal of Anthropological Research} 33, 452-473 (1977).
}

\keyword{datasets}



